remove invalid test, add auth to tests
