+++ src/backend/pipeline/test/unit/conftest.py
@@ -1,5 +1,6 @@
 import os
 import pytest
+from ...auth import init_firebase
 
 @pytest.fixture(scope="module", autouse=True)
 def set_directory():
@@ -8,3 +9,8 @@ def set_directory():
     abspath = os.path.abspath(__file__)
     dname = os.path.dirname(abspath)
     os.chdir(dname)
+
+@pytest.fixture(scope="session", autouse=True)
+def init_cloud_connections():
+    """Initializes the firebase connection."""
+    init_firebase()
